requirements.txt
# AI Trading Agent - Indodax Platform
# Core
ccxt
pandas==2.2.3
numpy==2.2.2

# Technical Analysis
ta==0.11.0

# HTTP & API
requests==2.32.3

# Configuration
python-dotenv==1.0.1

# Scheduling
schedule==1.2.2

# CLI Display
tabulate==0.9.0
colorama==0.4.6

# Testing
pytest==8.3.4
